Properties moved from iidm to stdcxx namespace (#110)
